tetragon: Move execve argument reader to header
As a preparation for execve code sharing moving read_args to
bpf_execve_event.h header, no function change.
