+++ bpf/errmetrics/fileids.h
@@ -29,3 +29,4 @@ fileid__("bpf_cgroup_events.h", 21)
 fileid__("bpf_fork.c", 22)
 fileid__("policy_filter.h", 23)
 fileid__("policy_stats.h", 24)
+fileid__("bpf_execve_event.h", 25)
